Adds Puppeteer support to frontend
This CL adds Puppeteer as frontend dependency. There are no call sites
that use this, and in this first CL only the Puppeteer Connection class
is exposed for use. As Puppeteer is agnostified to its environment, more
of it can be made available for use within the DevTools frontend
codebase.

Note: switching off the presubmit and win64 checks because it seems to
fail due to http://crbug.com/1068198. All other checks passed on prior
runs.

No-Presubmit: true
Cq-Exclude-Trybots: luci.devtools-frontend.try:devtools_frontend_win64_rel
Bug: 1107392
Change-Id: Ie29907af389eddb2e3a7bd260b64237529a9aeba