Adds Puppeteer support to frontend
This CL adds Puppeteer as frontend dependency. There are no call sites
that use this, and in this first CL only the Puppeteer Connection class
is exposed for use. As Puppeteer is agnostified to its environment, more
of it can be made available for use within the DevTools frontend
codebase.

Note: switching off the presubmit check because it seems to fail due to
http://crbug.com/1068198

No-Presubmit: true
Bug: 1107392
Change-Id: Ie29907af389eddb2e3a7bd260b64237529a9aeba