Adds Puppeteer support to frontend
This CL adds Puppeteer as frontend dependency. There are no call sites
that use this, and in this first CL only the Puppeteer Connection class
is exposed for use. As Puppeteer is agnostified to its environment, more
of it can be made available for use within the DevTools frontend
codebase.

Switching off trybots for this as:
1) the number of files passed to Windows bots for checking localization
causes a crash: namely that the command is too long. Previous patchsets
have passed on other bots, and on other non-localization tasks on win64.
2) This CL does not change any functionality, either, and is third party
code so does not warrant localization checks.

Bug: 1107392
Change-Id: Ie29907af389eddb2e3a7bd260b64237529a9aeba