Grid: Fix misplaced grid overlay labels on high DPI monitors
The recent work to support writing-modes in the grid overlay introduced
a regression in the way labels are positioned in the overlay.

What that change did was transform the canvas in order to draw lines
according to the current writing mode, and then use the canvas's matrix
to re-position all of the labels (the labels aren't drawn in the canvas
which is why this was a 2 steps process).

The problem is that the canvas is scaled according to the device pixel
ratio, and getting the transform matrix from it included this scaling.

Result: https://imgur.com/n7pGNpE.jpg

The labels positions don't need to be transformed with the scaling, only
with the matrix that corresponds to the writing-mode, since they are not
rendered in a canvas.

So the fix here is to scale the matrix with 1 / (scaling factor).

Bug: 1126348
Change-Id: I9df8d84f7fe6b01bb5099e46b6bf99122ef74fed