Implement GRD file check for devtools_pre_built
While we implemented this check for `devtools_module` and
`devtools_entrypoint`, we did not do so for `devtools_pre_built`.
This caused build failures yesterday when we forgot to add the
live directive of lit-html in the GRD files.

Therefore, implement the same check for `devtools_pre_built`. This
not only catches the live directive file missing, but also found out
that the imports for Acorn were using the wrong version. We use the
`.mjs` bundles from Acorn, not the `.js` versions.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Bug: 1126630
Change-Id: I54004efda6fb0ac50f894262f09fe2104d1b2f86