Load emulated_devices images from bundle instead of remote
This changes `emulated_devices` to load from bundled, via the
`importScriptPathPrefix` in Runtime, rather than the remoteBase.

For the build changes, we need to move the `_module.js` to non-remote,
have to update the GRD to include the files and copy the files to
the correct locations in the `resources/inspector/` and `gen/devtools/`.

All images are compressed with AVIF for the best compression rate, to
keep the impact on the Chromium binary size to a minimum.

This CL increases the bundle size of `resources.pak` from 5852KB to
6236KB.

R=jacktfranklin@chromium.org,aerotwist@chromium.org

Fixed: 1095614
Change-Id: If1bdceeda8c00e9981dc229fd7c81022bb6ae6f9