Fix adorner update race condition
The reason we saw duplicate adorners is because these
"updateStyleAdorners" calls are not atomic, and the async call happened
after we clear previous adorners, which mean that if other async calls
come back later to after the async call site, we will create duplicate
adorners.

The proper defensive fix would be to debounce "updateStyleAdorners" call
so that multiple calls in a very short interval only result in one call.
This will better prevent race conditions, as well as saving some
performance. This will happen in a follow-up CL because that involves
"commonize" debounce utility helper.

Bug: chromium:1126334
Change-Id: I4cfda254ce449a0b969526682b27b3d2f8ed3ab1