Grid: Give the grid overlay more contrast
When multiple grid overlays are enabled at once, it's kind of hard to
differentiate between grid containers on the page.

The feature crew agreed to make the container's border solid so it's
easier to see where a grid starts and ends.
Without this, the grid border is a dashed line, which makes it hard to
see if you're looking at a grid, or a cell inside a grid.

This change also increases the opacity of the extended grid lines a
little bit, so they can be seen more easily too.

Before: https://imgur.com/lhRa327.jpg
After: https://imgur.com/LOPurmV.jpg

Bug: 1047356
Change-Id: Iffe6e3aac06499b839f3f5cb08be1a5a1b9ae82a