Add keyboard shortcut editor experiment
This CL adds basic shortcut editing functionality to the shortcuts
settings, hidden behind the "enable shortcut editing" experiment.
With the experiment enabled, the shortcuts settings list will show an
edit button next to each action, including those that don't have
shortcuts by default.

The editor allows users to change, remove, or add
shortcuts, but does not allow modifier-only shortcuts or shortcuts that
are already mapped to another action. This second condition means that
all shortcuts are considered global for the purpose of conflict
detection, meaning that some user shortcuts that don't actually conflict
will be considered conflicts. For example, "h" to hide an element only
works in the Elements panel, but the current editor implementation will
not allow users to bind anything else to "h" as a result, even actions
that are specific to other panels. This may turn out to be overly
restrictive, but there isn't necessarily a good way to determine which
actions could conflict with eachother without going through all 90 and
defining exactly where they can happen--which may make sense to do in
the future, but is outside the scope of this CL.

This CL also adds a reset button to the shortcuts settings that resets
all shortcuts to their defaults. In the future, we'll allow users to
restore defaults on a per-shortcut basis.

Screenshot of editor: https://i.imgur.com/PTjOYG7.png

Design doc: https://docs.google.com/document/d/1oOPSWPxCHvMoBZ0Fw9jwFZt6gP4lrsrsl8DEAp-Hy7o/edit#

Bug: 174309
Change-Id: Idb1746e1a115d309a90d59daf26d021ff5337bd9