Fix background color stuck issue for Computed pane box model
This bug arises beause the metrics pane relied on "mouseover" event to
reset the background-color changes to the box model: when a "mouseover"
happens outside the box model but within the metrics pane, the colors
are reset. But when the console drawer is positioned so that it borders
the box model, the "mouseover" event won't happen when the mouse moves
from the box model to the console drawer, which means the last
"mouseover" event fired won't be considered as outside of box model yet
within metrics pane, and the colors won't be reset.

This CL adds the "mouseleave" event, which will be fired when the above
situation happens.

Bug: chromium:1124613
Change-Id: Ia408783c1c793e3076aa3dc2f52cd699d694f2f3