Grid: Avoid protocol errors when scrolling to grid containers
In a previous CL I added the ability to scroll a grid into view on the
page:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2404385

Turns out that or grid containers that are inside not rendered subtree,
this causes an error on the protocol (there's no exception int in the
frontend, and everything works fine though):
"Node does not have a layout object".

This error is returned by InspectorDOMAgent::scrollIntoViewIfNeeded.

To avoid this I'm making this feature consistent with how the
contextmenu/scroll into view feature works, by using the same method
on the DOMNode class instead of exposing a new method on the DOM domain.

Bug: 1109177
Change-Id: I68e483358338bf4c1f9f3cdfccbd93a414d82785