Fix loading of lighthouse_worker in release builds
In http://crrev.com/2401163 we changed the `_module.js` to be loaded
as ES modules. This broke loading remote modules such as the
lighthouse_worker, as the imports would only be resolved relative
to the current location. Since we serve almost all source from
`/bundled/`, lighthouse_worker would resolve to `/bundled/` as well.

Instead, we should force these remote modules to load from `/remote/`.
To do so, we have to use the `remoteBase` to load these `_module.js`
files.

Note that we are going to need do something similar once we migrate
remote modules to `modules` rather than `scripts`, where the
`entrypointFileName` will need to be resolved to `remoteBase` as well.

Also note that a local debug build of DevTools would have worked before,
as that adds the `debugFrontend` queryParam to load DevTools. As such,
it disables the remote lookup mechanism, to make working with remote
modules during development easier. This does mean that a release build
will "properly" exclude the `_module.js` file and therefore would show the
issue. To verify the issue is fixed, I have built Chromium without this
change and verified it is broken. After applying this patch, I have verified
this issue has been resolved.

R=aerotwist@chromium.org

Fixed: 1128890
Change-Id: I11ea2684332613f484a966506e73b8928f23c018