Extract inline CSS in inspector overlay
This CL extracts CSS for most tools. Since the order of unit test
execution is not guaranteed and something has changed with these changes
I updated a few tests that relied on global vars set by other tests.
Also I keep the grid styles inline for now because unit tests depend on
it.

Bug: 1100925
Change-Id: I883fe3f9e4e2fded3530cda037588f803632885b