Add unit test helpers for component DOM mutations
We recently discovered that the Breadcrumbs component is doing far more
work mutating DOM than it needs to. We therefore would like to put in
place some test helpers that will let us wrap tests in `withMutations`
or `withNoMutations` to assert on the expected mutations.

These helpers have been used to test the ElementBreadcrumbs, and the
improvements to those tests will come in a following CL. Splitting the
addition of the helpers into their own CL.

Change-Id: I3c8627bf242315db325f2f71c73d234947ba2aa3