Fix ElementsBreadcrumbs mutation bugs.
This CL fixes two bugs in the breadcrumbs that were causing too much
mutation:

1) Re-creating `NodeText` elements on each render.
2) Destroying the scroll buttons and then re-rendering them on every
   scroll.

The first is fixed by using `<devtools-node-text>` and deferring all
other rendering and instantiation concerns to Lit-Html/the browser. The
second is fixed by hiding and showing the buttons with CSS. We use the
new `withNoMutations` unit test helper to ensure we don't mutate and to
avoid mutations being inadvertently introduced again.

Bug: 1129881
Change-Id: I17fb517e2eca4761705d604ac34e6404a40b28f8