Remove autocompletion for box-shadow
In https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2232549
we added autocompletion for box-shadow, which replaced the three-dot autocompletion
feature. However, this has caused problems where typing <0> <space> would incorrectly
automplete to the full value. Thus, this creates user frustration when the user is
typing normally and wants to start a box-shadow with 0.

Instead, we should remove the space-completion on css-values. However, this requires
a longer investigation into what other parts are affected. To remove user frustration
in the short-term, we have decided that removing the auto-completion is the best
option for now. This also mimics the behavior of Chrome before, so is not strictly
a regression.

R=aerotwist@chromium.org

Fixed: 1126434
Change-Id: I8ced550f754d599b5e01f1db34bfeb64643bd403