Typecheck network/NetworkLogView.js with TypeScript
This file was using `Headers` as iterable. This class is iterable,
but TypeScript didn't like it. As it turns out, there are multiple
definitions of the `dom` and `webworker` APIs. To get the iterable
versions, we have to add them explicitly. Therefore, also update
our TypeScript configuration to add the iterable variants.

DISABLE_THIRD_PARTY_CHECK=TypeScript fix
R=szuend@chromium.org

Bug: 1011811
Change-Id: Iae32bbf25d55346081dc2e0ab4f4093b531f91a1