Don't display grid adorners on shadow roots
If an element in the Elements tree is a grid, it will get a grid adorner
but there's a bug where if that element contains a shadow root, then it
will also be getting the same adorner.

To avoid this, we skip document fragment nodes in the code.

Before: https://imgur.com/YU041Po.jpg
After: https://imgur.com/qaFdQle.jpg

Bug: 1129760
Change-Id: Ice4d9044b2a8a77a20d1db05ba4fb2acaa17775c