Add component hooks to conductor
These will be used in the tests for the style component. We are checking
these helpers/infra in before landing the component, to figure out how
to hook up the tests. I expect the `component_hooks.ts` to be refactored
away, but kept it now for ease of debugging/plumbing.

Bug:1126047
Change-Id: Ia01e16e66b22c353016313636f20f7d0745aac9e