Device emulation extra reloading
In device emulation, when rotating or selecting a different mode, there
is an extra reloading.

The reloading was introduced to make sure UI is refreshed to reflect
the new mode/orientation. But there are use cases where developers
hope to test the response with other events, not disrupted by the
reloading.

We decide to add a togglable setting for this behavior.

Bug: 1130853
Change-Id: I238c9eb369b1fe90c68f8ee658300e20adedc0cd