Adds support for loading text resources in Runtime
At the moment the Runtime loading code, fetchResource, assumes the same
root as the hosting module, i.e. if Lighthouse is on devtools/bundled,
it assumes that the loaded resource will also be on devtools/bundled.
This assumption does not hold true for Lighthouse loading its
localization files, which are held on devtools/remote. The Runtime
module currently provides a function for loading a binary resource, and
this CL adds an equivalent function for loading text resources, with the
expectation that modules like Lighthouse can use this to fetch remote
resources.

R=jacktfranklin@chromium.org

Change-Id: I723f8196904d62f3b024e662bfbbdb20742545d2