Ensure that data setters in components take an interface
It's clearer if a component's data setter is defined as:

  set data(data: Foo)

Rather than:

  set data(data: { x: string, y: number, ...})

And also has some advantages in that you can easily re-use the interface if you
define `get data`, and that you can use this type in Closure land if you need
to. So we are now enforcing at the component bridges level that each `set data`
follows this pattern.

This is part of a larger piece of work to ensure efficient DOM updates, because
exposing the interface like this lets us use it in a component's render code to
tell TypeScript how to type `.data=${{...}}` within lit-html.

Bug: 1129881
Change-Id: Ib8bbe5682acfed371d5afccc31727623167b1fdb