Fall back to utf-8 when previewing HTTP responses and charset is missing
When previewing responses in the network panel, an iframe is created and
loaded with a data-uri that contains the content data from the response.

When the responses came with the appropriate Content-Type header then
things are fine, as we use the charset from that header in building the
uri.
Sometimes, however, this header was not set by the server and the
preview contains garbled characters.

To attempt fixing this, we now fall back to utf-8 when the charset is
missing.

A test case where this happens is http://www.imomoe.in/
Before/After shot: https://imgur.com/JcemtSd.jpg

Bug: 1131397
Change-Id: Ie097731bd53de0e971f211a36e2b9d5630fe9552