New jump-to-var feature in the Styles pane
This implements a new feature called jump-to-var.

It allows users to quickly jump from a custom property used in a var()
CSS function to where this property is defined.
A simple link underline is displayed for those.

If the property used does not exist, then no underline appears.

The explainer for this feature can be found here:
https://docs.google.com/document/d/1LRwH_lIO-gqe8hWzVq8QGFsJ27eEl158nFsuPlLGfnU/edit#

Bug: 1124707
Change-Id: Ie9360f7dcace7621e4352fc4e6301e4c991b166e