Add singleton-instance lookup to sharedInstance
During the TypeScript migration, we are moving to reference singleton
instances, rather than the `sharedInstance` lookup in the runtime.
However, the sharedInstance wasn't aware of the singletons being
declared in (most notably) the panels. Therefore, it is possible
that developers migrate to the singleton pattern and consequently
cause two panels to be constructed.

To prevent this from happening, we update `sharedInstance` to lookup
the `instance` as declared on the type, if it exists. If it does,
it calls that function. This makes sure that singletons declared
won't break the Runtime/module.json setup that already exists.

For all panels/classes that perform the `sharedInstance` invocation
in their own declared `instance` method, we migrate to singletons
straight away. Otherwise, we end up with a StackOverflow where
both methods call each other.

R=aerotwist@chromium.org

Bug: 1131423, 1011811
Change-Id: Ice7d2f59114b9aaa8ddd8d24f7eefc7aa880ff4e