Turn on TypeScript import preservation
TypeScript's default behaviour if it sees an import such as:

```
import {X} from './foo.js'
```

Where X is a TypeScript type, _NOT_ a value, is to remove this import
when it compiles the code, because it thinks it's unused. This however
causes problems with our proposed solution for types in custom elements,
where we are importing:

```
import {NodeTextData} from './NodeText.js';
```

Here TypeScript will drop that import when it compiles, but that causes
the actual NodeText component to never be imported, and consequently
when we add <devtools-node-text> to the DOM, nothing renders as the
component hasn't been imported and registered.

The "preserve" setting of "importsNotUsedAsValues" will ensure that the
import above is kept and transformed into `import './NodeText.js';`.

If we do want to import types and have that dropped at compile time, we
should use the `import type {X}` syntax, which will be removed from the
compiled JavaScript.

DISABLE_THIRD_PARTY_CHECK=ts_library changes

Change-Id: Ifc0cd3aab7575c12ef7ee31b414b2d1abe894ed1