Fix bug with mutation expectations
If you run a unit test that expects some mutations:

```
withMutations([{ tagName: 'button', max: 1}], ...)
```

This test would fail if a button mutation was seen >1 times. However any
other mutations would be ignored. If your code also mutated some `p` tag
500 times, this would not be seen as a test failure.

This CL updates the mutation checker to find any mutations that were not
matched to a provided expectation from the developer, and errors for
them. We ignore #comment nodes because Lit-Html uses them to track DOM,
but it'll fail on any other DOM.

This change has been tested on the data-grid CL, which is WIP, where I
first discovered the problem.

Bug: 1129881
Change-Id: I61af7e92f6793f97d05b910504b3539274ea5444