Refactor away ad hoc treeElementSymbol field in DOMNode
Previously, several classes in the elements module set and access
an ad hoc field named `Symbol("treeElement")` on DOMNode. This is
dynamic and not part of DOMNode, and prevents proper typing. This CL
removes this behavior and make this DOMNode->TreeElement relationship
the responsibility of the ElementsTreeOutline, refactoring it away
from DOMNode.

Backend CL to temporarily disable tests: https://crrev.com/c/2439071

Change-Id: I3b3dcc9c3333bea1631fb9a781a39b7d48e9bbb7