Fix escape to close settings
In an earlier CL, I added !UI.isEditing() as a condition to UI.Dialog's
closing on esc behavior in order to allow the escape key to be handled
by the shortcut editor rather than the capturing listener on the dialog.

Since UI.isEditing() returns true when any input element is focused, it
prevents closing on tab when a checkbox is focused, which makes the
escape key basically unusable on the preferences or experiments tabs.
This CL removes the isEditing() check and instead just adds a mechanism
for the shortcut editor to disable closing on escape.

Bug: 1131395
Change-Id: I26154aebf35b2386ecf6c7e85b8857d01c4efecc