Network: Make the initiator column jump to the right position
The Network.Initiator CDP object will contain columnNumber wehn this CL
lands: https://chromium-review.googlesource.com/c/chromium/src/+/2442367
Before that, it only contained lineNumber.

This means that if, say, an HTML file is minified (all on 1 line) and
loads a CSS file, then the initiator's location will be on line 0 of
that HTML file. This is true, but incomplete. The CSS file might be
loaded by a <link> tag at column 10000 on that line.
So when we jump to the sources panel by clicking on the initiator link
and pretty-print the HTML file, we cannot link to the right position.

This change simply takes the new columnNumber info account and that's
enough to make the script formatter know where the initiator really is
in the pretty-printed file.

Bug: 1134064
Change-Id: Id8426e8856194e14afcf8e4ec80943e6b461b4f3