Fix heap snapshot test flakes on Mac bots
I was able to recreate this reliably on my Mac machine by upping the CPU
throttling. I added all the `steps` in such that I could see where it
failed and it always failed waiting for the retainer chain.

I then ran `it.repeat(50, ...)` with a timeout of 40,000ms and that
passed. I lowered it to 30,000 and about 90% of the tests passed, but
the odd one flaked. Therefore 35,000 seems the sweet spot.

I'd love to fix this "better" but there didn't seem to be an obvious
problem beyond this test taking time to run due to the nature of the
heap snapshot and how intensive it is to render and update.

Fixed: 1134093
Change-Id: Ibd160518b03067d7cda3077f2136f0ee1ff887a4