Make the screncast view tooltip more in line with the overlay tooltip
There are 2 places where overlay tooltips are built:
- in the real overlay that appears when using the Elements panel and
hovering over nodes,
- in the screencast overlay that appears when using the Elements panel
when debugging a remote device.

These tooltips look different, and that's fine for the most part, they
do not contain the same information.
But they both contain a preview of the hovered node (tag, id, class and
dimensions) and this preview should be similar across the 2 tooltips.

This change makes the text colors of the screencast view tooltip match
those of the normal tooltip.
It also fixes a bug where, in dark theme, the text inside the tooltip
was not visible easily.

Before/after shots: https://imgur.com/a/838Bwdo (after at the top)
For reference, the normal tooltip: https://imgur.com/a/V2Mt7hm

Bug: 1134087
Change-Id: I2c175e5abaa3c9c0f30478390fb1f8e8e19cd6ba