Fix conflicting scroll instructions in Breadcrumbs
The flakey tests on Windows exposed a bug where:

* We render and click the right button to scroll
* The code that ensures the active node is in view runs, and tries to
  scroll the active element into view.

These two actions would conflict with each other. On Linux/Mac and our
Windows bots, it seems the right scroll "wins" and the tests pass.

However, on Windows 10 the right scroll is overriden by scrolling the
active node into view, and therefore the test fails.

The fix we landed on is to flag if the user has manually scrolled the
crumbs, and gate the scrolling active node into view logic on that. If
the user manually scrolls, we effectively cede control and let them
control the scrolling. When we get an update (e.g. a new active selected
node), we take back control.

Change-Id: Iedf4cbe5bef652d16fc80a0e7c0b3b45a95ebbfa