Removed code that was blocking the clearing of continueToLocations
In this patch I removed the code in _onBlur in the debugger plugin that
was always returning true.  It appeared that the code was meant to
register whether the mouse was clicked in the debugger plugin or not,
however, since the event was blur it registered this as always being
true and kept the decorations on the continuteToLocations from being
cleared when the debugger plugin lost focus.

Bug: 811092
Change-Id: I0cb9a3d996d26c702c688e3cdf34a427d442cee1