timeline: Show JS even when we don't have a V8 invocation above it
generateJSFrameEvents() takes a profile trace after it has been
converted to contain JSSample events for the JS samples. For each
JSSample event, it creates JSFrame events that cover the span of the
sample.

Previously we discarded all JSSamples that were not nested inside of
a JS-invocation event like CallFunction. This is a problem when we start
profiling after the CallFunction has already begun, like in the case of
a worker that has a main loop where it waits on Atomics.wait.

This change shows all JS samples even if there is no such JS-invocation
event that they are nested in.

To do this we pretend like we have a top-level V8 invocation event, by
doing all of the things we would do in onStartEvent/onEndEvent. While
this fake event is activated (fakeJSInvocation), we treat all Instant
events as if they had a parent, which lets us keep collecting stacks.

When we see the next begin/end event in onStartEvent, we can end our
fake top-level parent.

Bug: 1127827
Change-Id: I9a638c5e2280938679cff88643829b993fc34e63