Use :host-context within component for dark mode
In DevTools you cannot use the `@media (prefers...)` to check for theme
because we also allow the user to set it within DevTools (e.g.
independently of their OS preference). We did this within the new
components by providing a helper class that you could interpolate into a
`style` tag within Lit-Html, but today I discovered `:host-context`
(https://developer.mozilla.org/en-US/docs/Web/CSS/:host-context()) which
enables us to use CSS only to solve this problem.

This CL updates ElementsBreadcrumbs to use it and rips out the JS hack
we'd used to work around it.

Bug: 1122511
Change-Id: I5a8c5cff4b77ee8b8c2cd35d8f79506f6b4c5b5e