Don't instantiate a SourceFormatter when modules are imported
This was causing problems when trying to write a unit test that
included a file which transitively included formatter.js, which was
instantiating a SourceFormatter, leading to a chain of unfortunate
events which relied on some other module already being loaded, which
it wasn't.

To avoid the dependency chain between the imports, expose a function
getSourceFormatter which lazily instantiates a shared instance.

Change-Id: I7c255186e53e131127ba00b7f719f2f1b34301c2