Track when custom properties are edited
Record a devtools action taken when users edit custom property values in
the Styles pane.
This is part 2 of the telemetry changes needed for the jump-to-var
feature. Part 1 was tracking when users click on the links, and was done
in this CL: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2451605/
This second part will allow us to compare how many users use those var
links with how many proceed to editing the value.

This depends on backend CL: to be added later.

Bug: 1124707
Change-Id: Ic4e783941c5bddaaff8adb47d7fd8fdab2ac27fb