Combine overlays into a single resource
This CL combines all overlays in a single entrypoint so that all
overlays are shipped in a single bundle. It introduces a new
call `setOverlay` that is made by the backend to specify which
overlay is active. Overlays implement install/uninstall methods
which allow enabling/disabling overlays without destroying the
page. Note that old entrypoints such tool_highlight.js can be removed
after this CL lands. The changes give the following advantages:

- we ship less code by shipping a single bundle (previously, different
bundles contained the same modules).
- we are one step closer to shipping the overlay page via CDP.
- we are able to more easily combine different overlays.

Backend CL:

Change-Id: I872a1d801879d90cc6d73adec5e88879216d1985